Send errors as exception events
Before this commit, the Remix instrumentation would report errors
with a series of attributes, containing whether an error took place,
the error message and its stack.

The recommended way to report errors in OpenTelemetry today is to
use an exception event and set the span status to error.

For backwards compatibility, a `legacyErrorAttributes` configuration
option is added, which re-enables attribute-based error reporting.