feat(remix): Add `http.route` to request handler spans (#41)
Even when a route does not have a loader or an action, it can still be useful to know the route that it matched on. This commit adds an `http.route` attribute to the request handler spans.

BREAKING CHANGE: The `remix.request` span name now appends the route path. Use the `code.function`: `requestHandler` attribute as a stable way to find the requestHandler span.